+++ test/presentation/pages/popular_movies_page_test.mocks.dart
@@ -9,7 +9,7 @@ import 'dart:ui' as _i7;
 import 'package:ditonton/common/state_enum.dart' as _i4;
 import 'package:ditonton/domain/entities/movie/movie.dart' as _i5;
 import 'package:ditonton/domain/usecases/movie/get_popular_movies.dart' as _i2;
-import 'package:ditonton/presentation/provider/popular_movies_notifier.dart'
+import 'package:ditonton/presentation/provider/movies/popular_movies_notifier.dart'
     as _i3;
 import 'package:mockito/mockito.dart' as _i1;
