+++ test/presentation/pages/top_rated_movies_page_test.mocks.dart
@@ -10,7 +10,7 @@ import 'package:ditonton/common/state_enum.dart' as _i4;
 import 'package:ditonton/domain/entities/movie/movie.dart' as _i5;
 import 'package:ditonton/domain/usecases/movie/get_top_rated_movies.dart'
     as _i2;
-import 'package:ditonton/presentation/provider/top_rated_movies_notifier.dart'
+import 'package:ditonton/presentation/provider/movies/top_rated_movies_notifier.dart'
     as _i3;
 import 'package:mockito/mockito.dart' as _i1;
